Import ChooserFactory classes in telepath endpoint files, rather than relying on them existing in the global scope
