+++ client/src/entrypoints/admin/page-chooser-telepath.js
@@ -1,4 +1,3 @@
-window.telepath.register(
-  'wagtail.widgets.PageChooser',
-  window.PageChooserFactory,
-);
+import { PageChooserFactory } from '../../components/ChooserWidget/PageChooserWidget';
+
+window.telepath.register('wagtail.widgets.PageChooser', PageChooserFactory);
