+++ client/src/entrypoints/documents/document-chooser-telepath.js
@@ -1,4 +1,6 @@
+import { DocumentChooserFactory } from '../../components/ChooserWidget/DocumentChooserWidget';
+
 window.telepath.register(
   'wagtail.documents.widgets.DocumentChooser',
-  window.DocumentChooserFactory,
+  DocumentChooserFactory,
 );
