+++ client/src/entrypoints/images/image-chooser-telepath.js
@@ -1,4 +1,6 @@
+import { ImageChooserFactory } from '../../components/ChooserWidget/ImageChooserWidget';
+
 window.telepath.register(
   'wagtail.images.widgets.ImageChooser',
-  window.ImageChooserFactory,
+  ImageChooserFactory,
 );
