+++ client/src/entrypoints/snippets/snippet-chooser-telepath.js
@@ -1,4 +1,6 @@
+import { SnippetChooserFactory } from '../../components/ChooserWidget/SnippetChooserWidget';
+
 window.telepath.register(
   'wagtail.snippets.widgets.SnippetChooser',
-  window.SnippetChooserFactory,
+  SnippetChooserFactory,
 );
