+++ wagtail/documents/views/chooser.py
@@ -172,7 +172,6 @@ class DocumentChooserAdapter(BaseChooserAdapter):
         return forms.Media(
             js=[
                 versioned_static("wagtaildocs/js/document-chooser-modal.js"),
-                versioned_static("wagtaildocs/js/document-chooser.js"),
                 versioned_static("wagtaildocs/js/document-chooser-telepath.js"),
             ]
         )
