+++ wagtail/images/widgets.py
@@ -56,7 +56,6 @@ class ImageChooserAdapter(BaseChooserAdapter):
         return forms.Media(
             js=[
                 versioned_static("wagtailimages/js/image-chooser-modal.js"),
-                versioned_static("wagtailimages/js/image-chooser.js"),
                 versioned_static("wagtailimages/js/image-chooser-telepath.js"),
             ]
         )
