+++ wagtail/snippets/widgets.py
@@ -48,7 +48,6 @@ class AdminSnippetChooser(BaseChooser):
     def media(self):
         return forms.Media(
             js=[
-                versioned_static("wagtailadmin/js/chooser-modal.js"),
                 versioned_static("wagtailsnippets/js/snippet-chooser.js"),
             ]
         )
@@ -61,8 +60,6 @@ class SnippetChooserAdapter(BaseChooserAdapter):
     def media(self):
         return forms.Media(
             js=[
-                versioned_static("wagtailadmin/js/chooser-modal.js"),
-                versioned_static("wagtailsnippets/js/snippet-chooser.js"),
                 versioned_static("wagtailsnippets/js/snippet-chooser-telepath.js"),
             ]
         )
